Fixed JAL instruction in corev-dv fence.i test
JAL takes symbol, not a number-immediate. In the test,
linker failures occurred in cases when the store
fencei exec stream randomization inserted a JAL instruction.

The instruction in question is supposed to never execute,
so the jump symbol is set to be at the beginning of the stream,
thus intentionally creating an inifite loop that will never
be entered.

Signed-off-by: Henrik Fegran <Henrik.Fegran@silabs.com>